scripts/backfill/requirements.txt
# Python dependencies for Sagitine Backfill Pipeline

# .msg file parsing
msgextract>=0.2.0

# Date parsing
python-dateutil>=2.8.0

# Optional: LLM Classification (uncomment to enable)
# anthropic>=0.18.0
